backend/requirements.txt
tonsdk
SQLAlchemy
tqdm
aiogram
requests
aiohttp
python-dotenv
